refactor(mysql): move SQL classification below app
