+++ src/app/policy/sql/mod.rs
@@ -1,6 +1,4 @@
 pub mod lexer;
-pub mod mysql_export;
-pub mod mysql_statement;
 pub mod result_query;
 pub mod sqlite_explain;
 pub mod sqlite_export;
