+++ src/domain/lib.rs
@@ -8,6 +8,7 @@ pub mod explain_plan;
 pub mod foreign_key;
 pub mod index;
 pub mod metadata;
+pub mod mysql_sql;
 pub mod query_history;
 pub mod query_result;
 pub mod rls;
